lib: Factor out /etc/os-release parsing
Move storaged's parser (which is really nice) into pkg/lib/, so that we
can use it from other places.
